requirements.txt
backports.tarfile==1.2.0
faiss-cpu==1.12.0
fastapi==0.117.1
fitz==0.0.1.dev2
frontend==0.0.3
importlib-metadata==8.0.0
ipykernel==6.30.1
jaraco.collections==5.1.0
ollama==0.5.4
pip-chill==1.0.3
pipreqs==0.4.13
polars==1.33.1
pymupdf==1.26.4
seaborn==0.13.2
sentence-transformers==5.1.0
tomli==2.0.1
